+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Please see each crate's change log below:
 
 - [`anyspawn`](./crates/anyspawn/CHANGELOG.md)
 - [`anyspawn_azure`](./crates/anyspawn_azure/CHANGELOG.md)
+- [`arty`](./crates/arty/CHANGELOG.md)
+- [`arty_io_core`](./crates/arty_io_core/CHANGELOG.md)
 - [`bytesbuf`](./crates/bytesbuf/CHANGELOG.md)
 - [`bytesbuf_io`](./crates/bytesbuf_io/CHANGELOG.md)
 - [`cachet`](./crates/cachet/CHANGELOG.md)
@@ -26,6 +28,7 @@ Please see each crate's change log below:
 - [`http_path_template`](./crates/http_path_template/CHANGELOG.md)
 - [`internity`](./crates/internity/CHANGELOG.md)
 - [`layered`](./crates/layered/CHANGELOG.md)
+- [`metabench`](./crates/metabench/CHANGELOG.md)
 - [`multitude`](./crates/multitude/CHANGELOG.md)
 - [`observed_macros_impl`](./crates/observed_macros_impl/CHANGELOG.md)
 - [`ohno`](./crates/ohno/CHANGELOG.md)